Fix build for gccgo.
gccgo requires uintptr types to be explicitly type converted back to a
uintptr before they can be used as an unsafe.Pointer.

Fixes #20.

+++ gdk/gdk.go
@@ -84,7 +84,7 @@ type Atom uintptr
 
 // Native() returns the underlying GdkAtom.
 func (v Atom) Native() C.GdkAtom {
-	return C.toGdkAtom(unsafe.Pointer(v))
+	return C.toGdkAtom(unsafe.Pointer(uintptr(v)))
 }
 
 /*
